requirements.txt
pinecone-haystack
haystack-ai
fastapi
uvicorn
python-dotenv
pathlib
pypdf
sentence-transformers>=2.2.0
python-multipart
-e .